7.商品的修改，删除，上下架

+++ e3-manager/e3-manager-dao/src/main/java/com/chinasofti/framework/BaseDao.java
@@ -13,9 +13,9 @@ public interface BaseDao<T>{
 
 	void update(T t);
 
-	void delete(Integer id);
+	void delete(long id);
 
-	T query(Integer id);
+	T query(long id);
 
 	List<T>queryByPageBean(T t, int firstResult, int maxResults, Object... params);
